requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    make upgrade
#
asgiref==3.6.0
    # via
    #   -r requirements/base.txt
    #   django
certifi==2023.5.7
    # via
    #   -r requirements/base.txt
    #   requests
cffi==1.15.1
    # via
    #   -r requirements/base.txt
    #   cryptography
    #   pynacl
charset-normalizer==3.1.0
    # via
    #   -r requirements/base.txt
    #   requests
click==8.1.3
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
code-annotations==1.3.0
    # via -r requirements/test.in
coverage[toml]==7.2.5
    # via pytest-cov
cryptography==40.0.2
    # via
    #   -r requirements/base.txt
    #   pyjwt
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-crum
    #   django-model-utils
    #   djangorestframework
    #   drf-jwt
    #   edx-django-utils
    #   edx-drf-extensions
    #   edx-organizations
django-crum==0.7.9
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
django-model-utils==4.3.1
    # via
    #   -r requirements/base.txt
    #   edx-organizations
django-object-actions==4.1.0
    # via -r requirements/base.txt
django-simple-history==3.0.0
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   edx-organizations
django-waffle==3.0.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
    #   edx-drf-extensions
djangorestframework==3.14.0
    # via
    #   -r requirements/base.txt
    #   drf-jwt
    #   edx-drf-extensions
    #   edx-organizations
drf-jwt==1.19.2
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
edx-django-utils==5.4.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
edx-drf-extensions==8.7.0
    # via
    #   -r requirements/base.txt
    #   edx-organizations
edx-opaque-keys==2.3.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
    #   edx-organizations
edx-organizations==6.11.1
    # via -r requirements/base.txt
exceptiongroup==1.1.1
    # via pytest
factory-boy==3.2.1
    # via -r requirements/test.in
faker==18.7.0
    # via factory-boy
freezegun==1.2.2
    # via -r requirements/test.in
future==0.18.3
    # via
    #   -r requirements/base.txt
    #   pyjwkest
idna==3.4
    # via
    #   -r requirements/base.txt
    #   requests
iniconfig==2.0.0
    # via pytest
jinja2==3.1.2
    # via code-annotations
markupsafe==2.1.2
    # via jinja2
newrelic==8.8.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
packaging==23.1
    # via pytest
pbr==5.11.1
    # via
    #   -r requirements/base.txt
    #   stevedore
pillow==9.5.0
    # via
    #   -r requirements/base.txt
    #   edx-organizations
pluggy==1.0.0
    # via pytest
psutil==5.9.5
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pycparser==2.21
    # via
    #   -r requirements/base.txt
    #   cffi
pycryptodomex==3.17
    # via
    #   -r requirements/base.txt
    #   pyjwkest
pyjwkest==1.4.2
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
pyjwt[crypto]==2.7.0
    # via
    #   -r requirements/base.txt
    #   drf-jwt
    #   edx-drf-extensions
pymongo==3.13.0
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
pynacl==1.5.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pytest==7.3.1
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==4.0.0
    # via -r requirements/test.in
pytest-django==4.5.2
    # via -r requirements/test.in
python-dateutil==2.8.2
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
    #   faker
    #   freezegun
python-slugify==8.0.1
    # via code-annotations
pytz==2023.3
    # via
    #   -r requirements/base.txt
    #   django
    #   djangorestframework
pyyaml==6.0
    # via code-annotations
requests==2.30.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
    #   pyjwkest
semantic-version==2.10.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
six==1.16.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
    #   pyjwkest
    #   python-dateutil
sqlparse==0.4.4
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.0.0
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
    #   edx-opaque-keys
text-unidecode==1.3
    # via python-slugify
tomli==2.0.1
    # via
    #   coverage
    #   pytest
urllib3==2.0.2
    # via
    #   -r requirements/base.txt
    #   requests
